+++ src/main.lisp
@@ -10,17 +10,17 @@
 (in-package ultralisp/main)
 
 
-(defmain main ((upload "Add this option to upload data to S3."
-                       :flag t)
-               (projects-dir "A directory where sources will checked out."
-                             :default "build/sources/"
-                             :short nil)
-               (dist-dir "A directory where sources will checked out."
-                         :default "build/dist/")
-               (name "A distribution's name."
-                     :default "ultralisp")
-               (base-url "A base url from which a distribution will be served."
-                         :default "http://dist.ultralisp.org/"))
+(defmain (main) ((upload "Add this option to upload data to S3."
+                         :flag t)
+                 (projects-dir "A directory where sources will checked out."
+                               :default "build/sources/"
+                               :short nil)
+                 (dist-dir "A directory where sources will checked out."
+                           :default "build/dist/")
+                 (name "A distribution's name."
+                       :default "ultralisp")
+                 (base-url "A base url from which a distribution will be served."
+                           :default "http://dist.ultralisp.org/"))
   (build :projects-dir projects-dir
          :dist-dir dist-dir
          :name name